third_party/blink/web_tests/fast/frames/sandboxed-iframe-domain-expected.txt

CONSOLE ERROR: Uncaught SecurityError: Failed to set the 'domain' property on 'Document': Assignment is forbidden for sandboxed iframes.
